Add Source File Directory, env and Server Script
The source files are contained within /src directory

We add a .env file for managing environment variables.
To allow collaboration, a sample of the environment variables used
are contained in .env.sample

We start our server by adding codes to app.ts inside the src directory.

This file is meant to get us started with building the features of
our application