Add Postgres & Redis Configuration Settings
Configuration for access to either of Postgres or Redis is initialized
in the Config directory.

The actual connection code is added in loaders directory.

This makes it easier to use the connection across several files.

To detect when something happens while trying to connect, changes
were made to our logging utility.