Add Basic Authentication Middleware
For any request to /inbound/sms or /outbound/sms , the request header
needs a basic authentication credential.

To check and parse this credential, we implement a middleware that
will be used across the necessary endpoints.

Also, for the account and phone_number table, we defined respective
models to point to each of them.